+++ mlir/lib/Gradient/Utils/GradientShape.cpp
@@ -28,7 +28,7 @@ namespace gradient {
 /// argument function result.
 ///
 std::vector<Type> computeResultTypes(func::FuncOp callee,
-                                     const std::vector<uint64_t> &diffArgIndices)
+                                     const std::vector<size_t> &diffArgIndices)
 {
     std::vector<Type> gradResultTypes;
     FunctionType fnType = callee.getFunctionType();
@@ -104,7 +104,7 @@ std::vector<Type> computeQGradTypes(func::FuncOp callee)
 /// The non differentiable params are filtered out.
 ///
 std::vector<Type> computeBackpropTypes(func::FuncOp callee,
-                                       const std::vector<uint64_t> &diffArgIndices)
+                                       const std::vector<size_t> &diffArgIndices)
 {
     std::vector<Type> backpropResTypes;
     FunctionType fnType = callee.getFunctionType();
